add dragging for template elmements
also restructuring the page directive, the file has got a lot bigger
with all that is going in, and it should get even bigger when we start
displaying elements dynamically corresponding to their types.

corrected an issue with the highlighing system, which still has few bugs
when you drag things around. needs more tinkering with, if anything bad
arises

dragula-angular is a nightmare, scopes are a headache, don't whish this
on my enemy.